requirements.txt
click
spotipy
numpy
pandas
python-dotenv
